(refactor) refactoed the withFeatureFlip
