+++ src/applications/check-in/containers/withFeatureFlip.jsx
@@ -1,20 +1,26 @@
 import React, { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 
-import { makeSelectFeatureToggles } from '../../utils/selectors/feature-toggles';
+import { makeSelectFeatureToggles } from '../utils/selectors/feature-toggles';
 
-const withFeatureFlip = Component => {
+const withFeatureFlip = (Component, options) => {
+  const { isPreCheckIn } = options;
   return props => {
     const selectFeatureToggles = useMemo(makeSelectFeatureToggles, []);
     const featureToggles = useSelector(selectFeatureToggles);
-    const { isCheckInEnabled, isLoadingFeatureFlags } = featureToggles;
+    const {
+      isCheckInEnabled,
+      isLoadingFeatureFlags,
+      isPreCheckInEnabled,
+    } = featureToggles;
+    const appEnabled = isPreCheckIn ? isPreCheckInEnabled : isCheckInEnabled;
     if (isLoadingFeatureFlags) {
       return (
         <>
           <va-loading-indicator message="Loading your check in experience" />
         </>
       );
-    } else if (!isCheckInEnabled) {
+    } else if (!appEnabled) {
       window.location.replace('/');
       return <></>;
     } else {
