+++ src/applications/check-in/pre-check-in/routes.jsx
@@ -10,7 +10,7 @@ import Landing from './pages/Landing';
 import Error from './pages/Error';
 import { URLS } from '../utils/navigation';
 
-import withFeatureFlip from './containers/withFeatureFlip';
+import withFeatureFlip from '../containers/withFeatureFlip';
 import withAuthorization from './containers/withAuthorization';
 import withForm from '../containers/withForm';
 
@@ -90,7 +90,7 @@ const createRoutesWithStore = () => {
         return (
           <Route
             path={`/${route.path}`}
-            component={withFeatureFlip(component)}
+            component={withFeatureFlip(component, { isPreCheckIn: true })}
             key={i}
           />
         );
